+++ orville-postgresql-libpq/package.yaml
@@ -58,7 +58,6 @@ library:
     - Orville.PostgreSQL.Internal.ErrorDetailLevel
     - Orville.PostgreSQL.Internal.ExecutionResult
     - Orville.PostgreSQL.Internal.Insert
-    - Orville.PostgreSQL.Internal.OrvilleState
     - Orville.PostgreSQL.Internal.PgTextFormatValue
     - Orville.PostgreSQL.Internal.PgTime
     - Orville.PostgreSQL.Internal.RawSql
@@ -75,6 +74,11 @@ library:
     - Orville.PostgreSQL.Marshall.SqlMarshaller
     - Orville.PostgreSQL.Marshall.SqlType
     - Orville.PostgreSQL.Marshall.SyntheticField
+    - Orville.PostgreSQL.Monad
+    - Orville.PostgreSQL.Monad.HasOrvilleState
+    - Orville.PostgreSQL.Monad.MonadOrville
+    - Orville.PostgreSQL.Monad.Orville
+    - Orville.PostgreSQL.OrvilleState
     - Orville.PostgreSQL.Plan
     - Orville.PostgreSQL.Plan.Explanation
     - Orville.PostgreSQL.Plan.Many
