+++ orville-postgresql-libpq/scripts/gen-local-docs.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+set -e
+
+docker-compose run \
+  --no-deps --rm dev \
+  stack --stack-yaml stack-lts-17.0.yml \
+  haddock --force-dirty --no-haddock-deps --haddock-arguments --odir=local-docs
