+++ orville-postgresql-libpq/src/Orville/PostgreSQL/Monad.hs
@@ -1,0 +1,15 @@
+{-# OPTIONS_GHC -Wno-missing-import-lists #-}
+
+module Orville.PostgreSQL.Monad
+  ( module Orville.PostgreSQL.Monad.Orville,
+    module Orville.PostgreSQL.Monad.HasOrvilleState,
+    module Orville.PostgreSQL.Monad.MonadOrville,
+  )
+where
+
+-- Note: we list the re-exports explicity above to control the order that they
+-- appear in the generated haddock documentation.
+
+import Orville.PostgreSQL.Monad.HasOrvilleState
+import Orville.PostgreSQL.Monad.MonadOrville
+import Orville.PostgreSQL.Monad.Orville
