+++ orville-postgresql-libpq/src/Orville/PostgreSQL/OrvilleState.hs
@@ -1,7 +1,6 @@
-{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE RankNTypes #-}
 
-module Orville.PostgreSQL.Internal.OrvilleState
+module Orville.PostgreSQL.OrvilleState
   ( OrvilleState,
     newOrvilleState,
     resetOrvilleState,
@@ -15,7 +14,6 @@ module Orville.PostgreSQL.Internal.OrvilleState
     openTransactionEvent,
     rollbackTransactionEvent,
     transactionSuccessEvent,
-    HasOrvilleState (askOrvilleState, localOrvilleState),
     ConnectionState (NotConnected, Connected),
     ConnectedState (ConnectedState, connectedConnection, connectedTransaction),
     connectState,
@@ -34,8 +32,6 @@ module Orville.PostgreSQL.Internal.OrvilleState
   )
 where
 
-import Control.Monad.Trans.Class (lift)
-import Control.Monad.Trans.Reader (ReaderT, ask, local, mapReaderT)
 import qualified Data.Map.Strict as Map
 import Data.Pool (Pool)
 
@@ -113,67 +109,6 @@ addTransactionCallback newCallback state =
         originalCallback event
         newCallback event
    in state {_orvilleTransactionCallback = wrappedCallback}
-
-{- |
-  'HasOrvilleState' is the typeclass that Orville uses to access and manange
-  the connection pool and state tracking when it is being executed inside an
-  unknown Monad. It is a specialized version of the Reader interface so that it
-  can easily implemented by application Monads that already have a Reader
-  context and want to simply add 'OrvilleState' as an attribute to that
-  context, like so
-
-  @
-    data MyApplicationState =
-      MyApplicationState
-        { appConfig :: MyAppConfig
-        , appOrvilleState :: OrvilleState
-        }
-
-    newtype MyApplicationMonad a =
-      MyApplicationMonad (ReaderT MyApplicationState IO) a
-
-    instance HasOrvilleState MyApplicationMonad where
-      askOrvilleState =
-        MyApplicationMonad (asks appOrvilleState)
-
-      localOrvilleState f (MyApplicationMonad reader) =
-        MyApplicationMonad $
-          local
-            (\state -> state { appOrvilleState = f (appOrvilleState state))
-            reader
-  @
-
-  An instance for 'ReaderT OrvilleState m' is provided as a convenience in
-  the case that your application has no extra context to track.
--}
-class HasOrvilleState m where
-  {-
-    Fetches the current 'OrvilleState' from the host Monad context. The
-    equivalent of 'ask' for 'ReaderT OrvilleState'
-  -}
-  askOrvilleState :: m OrvilleState
-
-  {-
-    Applies a modification to the 'OrvilleState' that is local to the given
-    monad operation. Calls to 'askOrvilleState' made within the 'm a' provided
-    must return the modified state. The modified state must only apply to
-    the given 'm a' and not persisted beyond it. The equivalent of 'local'
-    for 'ReaderT OrvilleState'
-  -}
-  localOrvilleState ::
-    -- | The function to modify the 'OrvilleState'
-    (OrvilleState -> OrvilleState) ->
-    -- | The monad operation to execute with the modified state
-    m a ->
-    m a
-
-instance Monad m => HasOrvilleState (ReaderT OrvilleState m) where
-  askOrvilleState = ask
-  localOrvilleState = local
-
-instance {-# OVERLAPS #-} (Monad m, HasOrvilleState m) => HasOrvilleState (ReaderT r m) where
-  askOrvilleState = lift askOrvilleState
-  localOrvilleState f = mapReaderT (localOrvilleState f)
 
 {- |
   Creates a appropriate initial 'OrvilleState' that will use the connection
